Code cleanup, made apt autoinstall flatpak
The code has been cleaned up a little, and more comments have been added so others can better understand the code. Apt will have the -y flag on its operations on the snapdpurge() function as the user will have already agreed to purge snapd and install Flatpak in the splash screen, so the code will now be able to automatically execute.